refactor: Separate select-category tab
